% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11.R
\name{BayesFates1}
\alias{BayesFates1}
\title{Mean}
\usage{
BayesFates1(
  Y1 = NULL,
  Y2 = NULL,
  Y3 = NULL,
  Y4 = NULL,
  b1_mu0_s_mu = NULL,
  b1_mu0_s_sd = NULL,
  b2_mu0_s_mu = NULL,
  b2_mu0_s_sd = NULL,
  b3_mu0_s_mu = NULL,
  b3_mu0_s_sd = NULL,
  b4_mu0_s_mu = NULL,
  b4_mu0_s_sd = NULL,
  b1_mu0_t_mu = NULL,
  b1_mu0_t_sd = NULL,
  b2_mu0_t_mu = NULL,
  b2_mu0_t_sd = NULL,
  b3_mu0_t_mu = NULL,
  b3_mu0_t_sd = NULL,
  b4_mu0_t_mu = NULL,
  b4_mu0_t_sd = NULL,
  b1_k_s_mu = NULL,
  b1_k_s_sd = NULL,
  b1_k_t_mu = NULL,
  b1_k_t_sd = NULL,
  b2_k_s_mu = NULL,
  b2_k_s_sd = NULL,
  b2_k_t_mu = NULL,
  b2_k_t_sd = NULL,
  b3_k_s_mu = NULL,
  b3_k_s_sd = NULL,
  b3_k_t_mu = NULL,
  b3_k_t_sd = NULL,
  b4_k_s_mu = NULL,
  b4_k_s_sd = NULL,
  b4_k_t_mu = NULL,
  b4_k_t_sd = NULL,
  lower_k_s = NULL,
  lower_k_t = NULL,
  lower_t0 = 0.2,
  upper_t0 = 0.8,
  upper_b1_t0 = 0.8,
  lower_b2_t0 = 0.2,
  alpha = 0.1,
  eplison = 0.04,
  lower_mu0_s = NULL,
  lower_mu0_t = NULL,
  upper_mu0_s = NULL,
  upper_mu0_t = NULL,
  student_df = 10,
  n_branch,
  gene_pairs = list(),
  cores = getOption("mc.cores", 1L),
  ...
)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
