% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTraj.R
\name{BayesTraj}
\alias{BayesTraj}
\title{BayesTraj}
\usage{
BayesTraj(
  Y1 = NULL,
  Y2 = NULL,
  Y3 = NULL,
  Y4 = NULL,
  n_branch,
  b1_mu0_s_mu = NULL,
  b1_mu0_s_sd = NULL,
  b2_mu0_s_mu = NULL,
  b2_mu0_s_sd = NULL,
  b3_mu0_s_mu = NULL,
  b3_mu0_s_sd = NULL,
  b4_mu0_s_mu = NULL,
  b4_mu0_s_sd = NULL,
  b1_mu0_t_mu = NULL,
  b1_mu0_t_sd = NULL,
  b2_mu0_t_mu = NULL,
  b2_mu0_t_sd = NULL,
  b3_mu0_t_mu = NULL,
  b3_mu0_t_sd = NULL,
  b4_mu0_t_mu = NULL,
  b4_mu0_t_sd = NULL,
  b1_k_s_mu = NULL,
  b1_k_s_sd = NULL,
  b1_k_t_mu = NULL,
  b1_k_t_sd = NULL,
  b2_k_s_mu = NULL,
  b2_k_s_sd = NULL,
  b2_k_t_mu = NULL,
  b2_k_t_sd = NULL,
  b3_k_s_mu = NULL,
  b3_k_s_sd = NULL,
  b3_k_t_mu = NULL,
  b3_k_t_sd = NULL,
  b4_k_s_mu = NULL,
  b4_k_s_sd = NULL,
  b4_k_t_mu = NULL,
  b4_k_t_sd = NULL,
  lower_k_s = NULL,
  lower_k_t = NULL,
  lower_t0 = 0.2,
  upper_t0 = 0.8,
  upper_b1_t0 = 0.8,
  lower_b2_t0 = 0.2,
  alpha = 0.1,
  eplison = 0.04,
  lower_mu0_s = NULL,
  lower_mu0_t = NULL,
  upper_mu0_s = NULL,
  upper_mu0_t = NULL,
  cores = getOption("mc.cores", 1L),
  ...
)
}
\arguments{
\item{Y1}{description}

\item{Y2}{description}

\item{Y3}{description}

\item{Y4}{description}

\item{n_branch}{number of branches}

\item{b1_mu0_s_mu}{description}

\item{b1_mu0_s_sd}{description}

\item{b2_mu0_s_mu}{description#'}

\item{b2_mu0_s_sd}{description}

\item{b3_mu0_s_mu}{description}

\item{b3_mu0_s_sd}{description}

\item{b4_mu0_s_mu}{description}

\item{b4_mu0_s_sd}{description}

\item{b1_mu0_t_mu}{description}

\item{b1_k_s_sd}{description}

\item{b1_k_t_mu}{description}

\item{b1_k_t_sd}{description}

\item{b2_k_s_mu}{description}

\item{b2_k_s_sd}{description}

\item{b2_k_t_mu}{description}

\item{b2_k_t_sd}{description}

\item{b3_k_s_mu}{description}

\item{b3_k_s_sd}{description}

\item{b3_k_t_mu}{description}

\item{b3_k_t_sd}{description}

\item{b4_k_s_mu}{description}

\item{b4_k_s_sd}{description}

\item{b4_k_t_mu}{description}

\item{b4_k_t_sd}{description}

\item{lower_k_s}{description}

\item{lower_k_t}{description}

\item{lower_t0}{description}

\item{upper_t0}{description}

\item{upper_b1_t0}{description}

\item{lower_b2_t0}{description}

\item{alpha}{description}

\item{eplison}{description}

\item{lower_mu0_s}{description}

\item{lower_mu0_t}{description}

\item{upper_mu0_s}{description}

\item{upper_mu0_t}{description}

\item{cores}{escription}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
