% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_calc.R
\name{BIC_calc}
\alias{BIC_calc}
\title{BIC_calc}
\usage{
BIC_calc(X, y, pi, beta, sigma, id, z)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{pi:}{Vector, group probabilities}

\item{beta:}{Matrix, likelihood coefficients}

\item{sigma:}{Float, likelihood variance}

\item{id:}{Vector, id for each observation}

\item{z:}{Matrix, K x dim(X)[2] matrix, indicating which variables to include in each group.}
}
\description{
Computes BIC for latent trajectory model
}
