% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_calc.R
\name{BIC_calc_dual}
\alias{BIC_calc_dual}
\title{BIC_calc_dual}
\usage{
BIC_calc_dual(X1, X2, y1, y2, pi1, pi1_2, beta1, beta2, sigma1, sigma2,
  id1, id2, z1, z2, constrain = FALSE)
}
\arguments{
\item{X1:}{Matrix, design matrix group 1}

\item{X2:}{Matrix, design matrix group 2}

\item{y1:}{Vector, outcomes group 1}

\item{y2:}{Vector, outcomes group 2}

\item{pi1:}{Vector, group probabilities group 1}

\item{pi1_2:}{Vector, transition probabilities from group 1 to group 2}

\item{beta1:}{Matrix, likelihood coefficients group 1}

\item{beta2:}{Matrix, likelihood coefficients group 2}

\item{sigma1:}{Float, likelihood variance group 1}

\item{sigma2:}{Float, likelihood variance group 2}

\item{id1:}{Vector, id for each observation group 1}

\item{id2:}{Vector, id for each observation group 2}

\item{z1:}{Matrix, K1 x dim(X1)[2] matrix, indicating which variables to include in each group.}

\item{z2:}{Matrix, K2 x dim(X2)[2] matrix, indicating which variables to include in each group.}
}
\description{
Compute BIC for dual latent trajectory model
}
