% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_branches}
\alias{calculate_branches}
\title{calculate_branches}
\usage{
calculate_branches(bt, limit_prob = 0.9)
}
\arguments{
\item{limit_prob}{概率大于多少认为是这个分支的细胞}

\item{bf}{Expression matrix: cells * genes}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
