% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_entropy}
\alias{calculate_entropy}
\title{calculate_entropy}
\usage{
calculate_entropy(bt)
}
\arguments{
\item{bf}{Expression matrix: cells * genes}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
