% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawbeta.R
\name{drawbeta}
\alias{drawbeta}
\title{drawbeta}
\usage{
drawbeta(X, Y, sigma, mu, Sigma)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{Y:}{Vector, outcomes}

\item{sigma:}{Float, variance of Y}

\item{mu:}{Vector, prior mean of beta}

\item{Sigma:}{Matrix, prior covariance of beta}
}
\description{
Draw beta from posterior distribution
}

