% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdegree.R
\name{drawdegree}
\alias{drawdegree}
\title{drawdegree}
\usage{
drawdegree(j, b, y, X, index, marg.lik.c)
}
\arguments{
\item{j:}{Integer, class flag}

\item{b:}{Integer, element of z[j,] to draw}

\item{y:}{Vector, outcomes}

\item{X:}{Matrix, design matrix}

\item{index:}{Vector: List of class assignments}

\item{marg.lik.c:}{Float: Current marginal likelihood}
}
\description{
Draw a component of d, the degree of the polynomial
}
