% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawgroup.R
\name{drawgroup}
\alias{drawgroup}
\title{drawgroup}
\usage{
drawgroup(X, y, N, id, pi, beta, sigma, K)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{N:}{Integer, number of units}

\item{id:}{Vector, tandem for each row of X and Y}

\item{pi:}{Vector, group probabilities}

\item{beta:}{Matrix, likelihood coefficients for}

\item{sigma:}{Float, likelihood variance}

\item{K:}{Number of groups}
}
\description{
Draw first series' groups from posterior distribution
}
