% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawgroup2.R
\name{drawgroup2}
\alias{drawgroup2}
\title{drawgroup2}
\usage{
drawgroup2(X, y, N, id, g1, pi1, pi1_2, beta, sigma, K)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{N:}{Integer, number of tandems}

\item{id:}{Vector, tandem for each row of X and Y}

\item{g1:}{Vector, group membership for series 1}

\item{pi1:}{Vector, group probabilities for series 1}

\item{pi1_2:}{Vector, transition probabilities for series 2}

\item{beta:}{Matrix, likelihood coefficients for series 1}

\item{sigma:}{Float or vector, likelihood variance}

\item{K:}{Number of groups in series 1}
}
\description{
Draw second series' groups from posterior distribution
}
