% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawgroup.R
\name{drawgroup_dual}
\alias{drawgroup_dual}
\title{drawgroup_dual}
\usage{
drawgroup_dual(X, y, N, id, g2, pi1, pi1_2, beta, sigma, K)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{N:}{Integer, number of pairs}

\item{id:}{Vector, tandem for each row of X and Y}

\item{g2:}{Vector, group membership for series 2}

\item{pi1:}{Vector, group probabilities for series 1}

\item{pi1_2:}{Vector, transition probabilities for series 2}

\item{beta:}{Matrix, likelihood coefficients for series 1}

\item{sigma:}{Float, likelihood variance}

\item{K:}{Number of groups in series 1}
}
\description{
Draw first series' groups from posterior distribution
}
