% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawmu.R
\name{drawmu}
\alias{drawmu}
\title{drawmu}
\usage{
drawmu(beta, Sigma, Lambda, K)
}
\arguments{
\item{beta:}{beta: Matrix, coefficients for outcome in each row}

\item{Sigma:}{Matrix, prior covariance of beta}

\item{Lambda:}{Matrix, prior covariance of mu}

\item{K:}{Number of groups}
}
\description{
Draw mu from posterior distribution
}
