% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpi.R
\name{drawpi}
\alias{drawpi}
\title{drawpi}
\usage{
drawpi(g, alpha, K)
}
\arguments{
\item{g:}{Vector, group memberships}

\item{alpha:}{Vector, parameters of Dirichlet prior}

\item{K:}{Integer, number of groups}
}
\description{
Draw pi from posterior distribution
}
