% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpoly.R
\name{drawpoly}
\alias{drawpoly}
\title{drawpoly}
\usage{
drawpoly(y, X, time_index, z)
}
\arguments{
\item{z:}{Matrix, current variable selection matrix}

\item{j:}{Integer, class flag}

\item{b:}{Integer, element of z[j,] to draw}

\item{y:}{Vector, outcomes}

\item{X:}{Matrix, design matrix}

\item{index:}{Vector: List of class assignments}

\item{marg.lik.c:}{Float: Current marginal likelihood}
}
\description{
Draw the polynomial degree for time
}
