% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawz.R
\name{drawz}
\alias{drawz}
\title{drawz}
\usage{
drawz(z, j, b, y, X, index, marg.lik.c)
}
\arguments{
\item{z:}{Matrix, current variable selection matrix}

\item{j:}{Integer, class flag}

\item{b:}{Integer, elemtn of z[j,] to draw}

\item{y:}{Vector, outcomes}

\item{X:}{Matrix, design matrix}

\item{index:}{Vector: List of class assignments}

\item{marg.lik.c:}{Float: Current marginal likelihood}
}
\description{
Draw a component of z, the vectod determining which variables are included in the model.
}
