% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawzeta.R
\name{drawzeta}
\alias{drawzeta}
\title{drawzeta}
\usage{
drawzeta(c, W, zeta, N, P, J)
}
\arguments{
\item{c:}{Vector, group memberships}

\item{W:}{Matrix, design matrix}

\item{zeta:}{Matrix, previous draw of mnlogit parameters}

\item{N:}{Integer, number of units}

\item{P:}{Integer, number of covariates}

\item{J:}{Integer, number of groups}
}
\description{
Draw group determination mnlogit parameters.
This is a slight modification of the BayesLogit mlogit.R function
}
