% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtraj_constrained.R
\name{dualtraj_constrained}
\alias{dualtraj_constrained}
\title{dualtraj_consrained}
\usage{
dualtraj_constrained(X1, X2, y1, y2, K, z, iterations, thin = 1,
  dispIter = 10, ll = FALSE)
}
\arguments{
\item{X1:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{X2:}{Matrix, design matrix for series 2. 1st column should be the id.}

\item{y1:}{Vector, outcomes for series 1}

\item{y2:}{Vector, outcomes for series 2}

\item{K:}{Integer, number of latent classes}

\item{z:}{Matrix, K x dim(X)[2] indicator matrix indicating which variables to inlcude in each group.}

\item{iterations:}{Integer, number of MCMC iterations}

\item{thin:}{Integer, store every 'thin' iteration}

\item{dispIter:}{Integer, frequency of printing the iteration number}

\item{ll:}{Boolean, Set to TRUE to display the maximum log-likelihood over all the draws.}
}
\description{
Estimate the dual trajectory model using a Gibbs Sampler
}
