% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtrajMS.R
\name{dualtrajMS}
\alias{dualtrajMS}
\title{dualtraj}
\usage{
dualtrajMS(X1, X2, y1, y2, K1, K2, iterations, thin = 1, dispIter = 10)
}
\arguments{
\item{X1:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{X2:}{Matrix, design matrix for series 2. 1st column should be the id.}

\item{y1:}{Vector, outcomes for series 1}

\item{y2:}{Vector, outcomes for series 2}

\item{K1:}{Integer, number of latent classes in series 1}

\item{K2:}{Integer, number of latent classes in series 2}

\item{iterations:}{Integer, number of MCMC iterations}

\item{thin:}{Integer, store every 'thin' iteration}

\item{dispIter:}{Integer, frequency of printing the iteration number}
}
\description{
Estimate the dual trajectory model using a Gibbs Sampler
}
