% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data2.R
\name{gen_data2}
\alias{gen_data2}
\title{gen_data}
\usage{
gen_data2(N, T1, T2, pi1, pi2, beta1, beta2, sigma1, sigma2)
}
\arguments{
\item{N:}{Integer, number of tandems}

\item{T1:}{Integer, number of time periods in series 1}

\item{T2:}{Integer, number of time periods in series 2}

\item{pi1:}{Vector, probability of being assigned to each group in series 1}

\item{pi2:}{Matrix, transition matrix. Row j is the transition probabilities for when the series 1 group is j}

\item{beta1:}{Matrix, coefficients for each group in series 1}

\item{beta2:}{Matrix, coefficients for each group in series 2}

\item{sigma1:}{Float, variance for outcomes in series 1}

\item{sigma2:}{Float, variance for outcomes in series 2}
}
\value{
List: \cr
   X1: Matrix, dataset for series 1 \cr
   X2: Matrix, dataset for series 2 \cr
   Y1: Vector, outcomes for series 1 \cr
   Y2: Vector, outcomes for series 2 \cr
}
\description{
Generate data for a dual trajectory model
}

