% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_traj.R
\name{gen_data_traj}
\alias{gen_data_traj}
\title{gen_data_traj}
\usage{
gen_data_traj(N, T1, pi, beta, sigma, zeta = matrix(), scale = FALSE)
}
\arguments{
\item{N:}{Integer, number of tandems}

\item{T1:}{Integer, number of time periods}

\item{pi1:}{Vector, probability of being assigned to each group}

\item{beta:}{Matrix, coefficients for each group}

\item{sigma:}{Float, variance for outcomes in series 1}

\item{scale:}{Boolean, TRUE to scale design matrix}
}
\value{
List: \cr
   X1: Matrix, dataset \cr
   Y1: Vector, outcomes \cr
}
\description{
Generate data for a single trajectory model
}
