% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_pseudotime}
\alias{infer_pseudotime}
\title{infer_pseudotime}
\usage{
infer_pseudotime(bt)
}
\arguments{
\item{fit}{Expression matrix: cells * genes}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
