% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infer_t0}
\alias{infer_t0}
\title{infer_t0}
\usage{
infer_t0(bt)
}
\arguments{
\item{bf}{Expression matrix: cells * genes}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
