% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{log_lik}
\alias{log_lik}
\title{log_lik}
\usage{
log_lik(X, y, pi, beta, sigma, id)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{pi:}{Vector, group probabilities}

\item{beta:}{Matrix, likelihood coefficients}

\item{sigma:}{Float, likelihood variance}

\item{id:}{Vector, id for each observation}
}
\description{
Compute log likelihood of data for latent trajectory model
}
