% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{log_lik_dual}
\alias{log_lik_dual}
\title{log_lik_dual}
\usage{
log_lik_dual(X1, X2, y1, y2, pi1, pi1_2, beta1, beta2, sigma1, sigma2, id1, id2)
}
\arguments{
\item{X1:}{Matrix, design matrix group 1}

\item{X2:}{Matrix, design matrix group 2}

\item{y1:}{Vector, outcomes group 1}

\item{y2:}{Vector, outcomes group 2}

\item{pi1:}{Vector, group probabilities group 1}

\item{pi1_2:}{Vector, transition probabilities from group 1 to group 2}

\item{beta1:}{Matrix, likelihood coefficients group 1}

\item{beta2:}{Matrix, likelihood coefficients group 2}

\item{sigma1:}{Float or vector, likelihood variance group 1}

\item{sigma2:}{Float or vector, likelihood variance group 2}

\item{id1:}{Vector, id for each observation group 1}

\item{id2:}{Vector, id for each observation group 2}
}
\description{
Compute log likelihood of data for dual latent trajectory model
}
