% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{log_lik_obs}
\alias{log_lik_obs}
\title{log_lik_obs}
\usage{
log_lik_obs(X, y, beta, sigma)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{y:}{Vector, outcomes}

\item{beta:}{Matrix, likelihood coefficients}

\item{sigma:}{Float or Vector, likelihood variance}
}
\description{
Compute log likelihood of each observation
}
