% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marglik.R
\name{marg_lik}
\alias{marg_lik}
\title{marg_lik}
\usage{
marg_lik(y, X)
}
\arguments{
\item{y:}{Vector, outcomes}

\item{X:}{Matrix, design matrix}
}
\description{
Calculate log marginal likelihood under unit-information Zellner's g-prior
}
