% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalized_expression_data}
\alias{normalized_expression_data}
\title{normalized_expression_data}
\usage{
normalized_expression_data(expression_data)
}
\arguments{
\item{expression_data}{Expression matrix: cells * genes}
}
\description{
A function to compute the mean of a vector
}
\examples{
mean(1:3)
\dontrun{ mean(1:1e99) }

}
