% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_dual}
\alias{summary_dual}
\title{summary_dual}
\usage{
summary_dual(model, X1, X2, y1, y2, z1, z2, burn)
}
\arguments{
\item{model:}{List, output from dualtraj or dualtrajMS}

\item{X1:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{X2:}{Matrix, design matrix for series 2. 1st column should be the id.}

\item{y1:}{Vector, outcomes for series 1}

\item{y2:}{Vector, outcomes for series 2}

\item{z1:}{Matrix, K1 x dim(X1)[2] indicator matrix indicating which variables to inlcude in each group.}

\item{z2:}{Matrix, K2 x dim(X2)[2] indicator matrix indicating which variables to inlcude in each group.}

\item{burn:}{float, fraction of draws to keep for post burn-in period.}
}
\description{
Summarize output of dual trajectory model
}
