% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_single}
\alias{summary_single}
\title{summary_single}
\usage{
summary_single(model, X, y, z, burn)
}
\arguments{
\item{model:}{List, output from dualtraj or dualtrajMS}

\item{X:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{y:}{Vector, outcomes for series 1}

\item{z:}{Matrix, K x dim(X)[2] indicator matrix indicating which variables to inlcude in each group.}

\item{burn:}{float, fraction of draws to keep for post burn-in period.}
}
\description{
Summarize output of single trajectory model
}
