% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_single_MS}
\alias{summary_single_MS}
\title{summary_single_MS}
\usage{
summary_single_MS(model, X, y, burn)
}
\arguments{
\item{model:}{List, output from dualtraj or dualtrajMS}

\item{X:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{y:}{Vector, outcomes for series 1}

\item{burn:}{float, fraction of draws to keep for post burn-in period.}
}
\description{
Summarize output of single trajectory model with Bayesian model averaging.
}
