% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawgroup2.R
\name{drawgroup2}
\alias{drawgroup2}
\title{drawbeta}
\usage{
drawgroup2(X, Y, N, id, g2, pi1, pi2, beta, sigma, K)
}
\arguments{
\item{X:}{Matrix, design matrix}

\item{Y:}{Vector, outcomes}

\item{N:}{Integer, number of tandems}

\item{id:}{Vector, tandem for each row of X and Y}

\item{g2:}{Vector, group membership for series 2}

\item{pi1:}{Vector, group probabilities for series 1}

\item{pi2:}{Vector, transition probabilities for series 2}

\item{beta:}{Matrix, likelihood coefficients for series 1}

\item{sigma:}{Float, likelihood variance}

\item{K:}{Number of groups in series 1}
}
\description{
Draw beta from posterior distribution
}

