% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawpoly.R
\name{drawpoly}
\alias{drawpoly}
\title{drawpoly}
\usage{
drawpoly(y, X, time_index, z)
}
\arguments{
\item{y:}{Vector, outcomes}

\item{X:}{Matrix, design matrix}

\item{time_index:}{Integer, column of X corresponding to time}

\item{z:}{Vector, current variable selection matrix}
}
\description{
Draw the polynomial degree for time
}
