% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualtraj.R
\name{dualtraj}
\alias{dualtraj}
\title{dualtraj}
\usage{
dualtraj(X1, X2, Y1, Y2, K1, K2, iterations)
}
\arguments{
\item{X1:}{Matrix, design matrix for series 1. 1st column should be the id.}

\item{X2:}{Matrix, design matrix for series 2. 1st column should be the id.}

\item{Y1:}{Vector, outcomes for series 1}

\item{Y2:}{Vector, outcomes for series 2}

\item{K1:}{Integer, number of latent classes in series 1}

\item{K2:}{Integer, number of latent classes in series 2}

\item{iterations:}{Integer, number of MCMC iterations}
}
\description{
Estimate the dual trajectory model using a Gibbs Sampler
}

